# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/markers/{id}/comments": {
      "get": {
        "operationId": "Social_listComments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "nextCursor from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "A marker's comments, oldest first",
        "tags": [
          "social"
        ]
      },
      "post": {
        "description": "Plain text, 1–1000 characters. Members, or anyone for a public trip. Not on markers of someone I have a block with (403). 30 per minute.",
        "operationId": "Social_createComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "A client-generated UUID, new per action and reused on its retries. For 24 hours a retry with the same key and body returns the first response (header `Idempotent-Replayed: true`); the same key with a different request answers 422 IDEMPOTENCY_KEY_REUSED, and while the first is still running 409 IDEMPOTENCY_KEY_IN_PROGRESS.",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 255
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Comment on a marker",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/comments/{id}": {
      "delete": {
        "description": "The author or the trip owner. Idempotent.",
        "operationId": "Social_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a comment",
        "tags": [
          "social"
        ]
      }
    },
    "/v1/likes/{targetType}/{targetId}": {
      "put": {
        "description": "Idempotent. Returns the new like state and count.",
        "operationId": "Social_like",
        "parameters": [
          {
            "name": "targetType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "enum": [
                "trip",
                "marker",
                "photo",
                "comment",
                "place"
              ]
            }
          },
          {
            "name": "targetId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Like a trip, marker, photo, comment or place",
        "tags": [
          "social"
        ]
      },
      "delete": {
        "description": "Idempotent.",
        "operationId": "Social_unlike",
        "parameters": [
          {
            "name": "targetType",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "enum": [
                "trip",
                "marker",
                "photo",
                "comment",
                "place"
              ]
            }
          },
          {
            "name": "targetId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LikeStateDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove my like",
        "tags": [
          "social"
        ]
      }
    },
...
  },
  "info": {
    "title": "Tripinly API",
    "description": "REST API for the Tripinly apps. Errors use `{ error: { code, message, details } }`; switch on `code`.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string",
            "example": "jonas.k"
          },
          "displayName": {
            "type": "string",
            "example": "Jonas"
          }
        },
        "required": [
          "id",
          "username",
          "displayName"
        ]
      },
...
      "CommentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "markerId": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "author": {
            "$ref": "#/components/schemas/UserSummaryDto"
          },
          "likeCount": {
            "type": "number"
          },
          "likedByMe": {
            "type": "boolean"
          },
          "canDelete": {
            "type": "boolean",
            "description": "I wrote it, or I own the trip"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "markerId",
          "body",
          "author",
          "likeCount",
          "likedByMe",
          "canDelete",
          "createdAt"
        ]
      },
      "CommentPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "minLength": 1,
            "maxLength": 1000
          }
        },
        "required": [
          "body"
        ]
      },
      "LikeStateDto": {
        "type": "object",
        "properties": {
          "targetType": {
            "type": "string",
            "enum": [
              "trip",
              "marker",
              "photo",
              "comment",
              "place"
            ]
          },
          "targetId": {
            "type": "string",
            "format": "uuid"
          },
          "liked": {
            "type": "boolean"
          },
          "likeCount": {
            "type": "number",
            "description": "The target's like count after the change (a place's popularity for places)"
          }
        },
        "required": [
          "targetType",
          "targetId",
          "liked",
          "likeCount"
        ]
      },
...
    }
  }
}